Atualização do backend - Claudio Eduardo

+++ codes/backend/app/Http/Controllers/AdminController.php
@@ -1,9 +1,10 @@
 <?php
 
 namespace App\Http\Controllers;
-use Illuminate\Support\Facades\Storage;
-use App\Models\Admin;
+
 use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Storage;
 
 class AdminController extends Controller
 {
@@ -12,13 +13,14 @@ class AdminController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function __construct(Admin $admin){
+    public function __construct(User $admin){
         $this->admin = $admin;
     }
 
     public function index()
     {
         $admin = $this->admin->all();
+        $admin = $admin->where('isAdmin', 1);
         return response()->json($admin, 200);
     }
 
@@ -40,16 +42,20 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate($this->admin->rules(), $this->admin->feedback());
+        $request->validate($this->admin->rulesAdmin(), $this->admin->feedback());
         $imagem = $request->file('imagem');
-        $imagem_urn = $imagem->store('imagens', 'public');
+        $imagem_urn = $imagem->store('imagens/admins', 'public');
 
         $admin = $this->admin->create([
-            'nome' => $request->nome, 
-            'sobrenome' => $request->sobrenome, 
-            'email' => $request->email, 
-            'password' => $request->password, 
-            'imagem' => $imagem_urn
+            'imagem' => $imagem_urn,
+            'name' => $request->name,
+            'sobrenome' => $request->sobrenome,
+            'email' => $request->email,
+            'password' => $request->password,
+            'usuario' => $request->usuario,
+            'telefone' => $request->telefone,
+
+            'isAdmin' => 1
         ]);
         return response()->json($admin);
     }
@@ -57,7 +63,7 @@ class AdminController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  Integer
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -66,16 +72,19 @@ class AdminController extends Controller
         if($admin == null){
             return response()->json(['erro' => 'recurso pesquisado não existe'], 404);
         }
+        if($admin['isAdmin'] == 0){
+            return response()->json(['erro' => 'recurso pesquisado não existe'], 404);
+        }
         return response()->json($admin, 200);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Admin  $admin
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Admin $admin)
+    public function edit($id)
     {
         //
     }
@@ -84,7 +93,7 @@ class AdminController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  Integer
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -93,12 +102,16 @@ class AdminController extends Controller
         if($admin == null){
             return response()->json(['erro' => 'impossível realizar a actualização. recurso não encontrado'], 404);
         }
+        if($admin['isAdmin'] == 0){
+            return response()->json(['erro' => 'recurso pesquisado não existe'], 404);
+
+        }
 
         if($request->method() == 'PATCH'){
 
             $regras = array();
 
-            foreach($admin->rules() as $input => $regra){
+            foreach($admin->rulesAdmin() as $input => $regra){
 
                 if(array_key_exists($input, $request->all())){
 
@@ -109,7 +122,7 @@ class AdminController extends Controller
             $request->validate($regras, $admin->feedback());
         }else{
 
-            $request->validate($admin->rules(), $admin->feedback());
+            $request->validate($admin->rulesAdmin(), $admin->feedback());
 
         }
 
@@ -118,23 +131,20 @@ class AdminController extends Controller
         }
 
         $imagem = $request->file('imagem');
-        $imagem_urn = $imagem->store('imagens', 'public');
+        $imagem_urn = $imagem->store('imagens/admins', 'public');
 
         
-        $admin->update([
-            'nome' => $request->nome, 
-            'sobrenome' => $request->sobrenome, 
-            'email' => $request->email, 
-            'password' => $request->password, 
-            'imagem' => $imagem_urn
-        ]);
+        $admin->fill($request->all());
+        $admin->imagem = $imagem_urn;
+        $admin->save();
+
         return response()->json($admin, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  Integer
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
@@ -144,10 +154,14 @@ class AdminController extends Controller
             return response()->json(['erro' => 'não existe o recurso que se pretende excluir'], 404);
         }
 
+        if($admin['isAdmin'] == 0){
+            return response()->json(['erro' => 'recurso pesquisado não existe'], 404);
+
+        }
+
         Storage::disk('public')->delete($admin->imagem);
 
         $admin->delete();
         return response()->json(['recurso excluido com sucesso']);
     }
-    
 }
